Renaming tests to IT in order to not encounter rest cancellable node client handler failures
